+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptor/sections/eventlisteners/DeploymentsEventListenersView.html
@@ -14,13 +14,16 @@
   ~ limitations under the License.
   -->
 
-<form>
-    <h3 data-i18n-key="Global"></h3>
+<form role="form">
+
+    <h3 data-field="title" data-i18n-key="EventListeners"></h3>
+
+    <br>
+
     <table class="table table-hover kie-tablelist">
         <thead>
         <tr class="kie-tablelist__row kie-tablelist__row--header">
             <th class="kie-tablelist__cell kie-tablelist__cell--header" data-i18n-key="Name"></th>
-            <th class="kie-tablelist__cell kie-tablelist__cell--header" data-i18n-key="Value"></th>
             <th class="kie-tablelist__cell kie-tablelist__cell--header" data-i18n-key="Resolver"></th>
             <th class="kie-tablelist__cell kie-tablelist__cell--header" data-i18n-key="Parameters"></th>
             <th class="kie-tablelist__cell kie-tablelist__cell--header kie-tablelist__cell--actions">
@@ -28,13 +31,13 @@
             </th>
         </tr>
         </thead>
-        <tbody data-field="globals">
+        <tbody data-field="event-listeners">
         </tbody>
 
     </table>
 
-    <button type="button" class="btn btn-link" data-field="add-global-button">
+    <button type="button" class="btn btn-link" data-field="add-event-listener-button">
         <span class="pficon pficon-add-circle-o"></span>
-        <span data-i18n-key="AddGlobal"></span>
+        <span data-i18n-key="AddEventListener"></span>
     </button>
 </form>
