+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptor/sections/general/DeploymentsGeneralSettingsView.html
@@ -18,6 +18,8 @@
 
     <h3 data-field="title" data-i18n-key="GeneralSettings"></h3>
 
+    <br>
+
     <div class="form-group">
         <label data-i18n-key="RuntimeStrategy"></label>
         <div data-field="runtime-strategies"></div>
