+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptor/sections/requiredroles/DeploymentsRequiredRolesView.html
@@ -14,8 +14,12 @@
   ~ limitations under the License.
   -->
 
-<form>
-    <h3 data-i18n-key="RequiredRoles"></h3>
+<form role="form">
+
+    <h3 data-field="title" data-i18n-key="RequiredRoles"></h3>
+
+    <br>
+
     <table class="table table-hover kie-tablelist">
         <thead>
         <tr class="kie-tablelist__row kie-tablelist__row--header">
